requirements.txt
django
django_heroku
djangorestframework
django-filter  
requests
psycopg2
pyyaml 
uritemplate
gunicorn
dj_database_url
markdown
whitenoise
python-dotenv
Pillow
cloudinary